# front/openapi.yaml
openapi: 3.1.0
info:
  title: FerrisKey API
  description: FerrisKey API
  contact:
    name: nathaelb
    email: pro.nathaelbonnal@gmail.com
  license:
    name: ""
  version: 0.1.0
paths:
...
  /realms/{realm_name}/files:
    get:
      tags:
        - file
      summary: List files in a realm
      operationId: list_files
      parameters:
        - name: realm_name
          in: path
          description: Realm name
          required: true
          schema:
            type: string
        - name: offset
          in: query
          description: "Pagination offset (default: 0)"
          required: false
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          description: "Pagination limit (default: 20, max: 100)"
          required: false
          schema:
            type: integer
            format: int64
        - name: mime_type
          in: query
          description: Filter by MIME type
          required: false
          schema:
            type: string
        - name: uploaded_by
          in: query
          description: Filter by uploader user ID
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Files listed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paginated_StoredObject"
        "400":
          description: Bad request
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Paginated_StoredObject:
      type: object
      required:
        - items
        - offset
        - limit
        - count
      properties:
        count:
          type: integer
          format: int64
        items:
          type: array
          items:
            type: object
            required:
              - id
              - realm_id
              - bucket
              - object_key
              - original_name
              - mime_type
              - size_bytes
              - checksum_sha256
              - metadata
              - uploaded_by
              - created_at
              - updated_at
            properties:
              bucket:
                type: string
              checksum_sha256:
                type: string
              created_at:
                type: string
                format: date-time
              created_by:
                type:
                  - string
                  - "null"
                format: uuid
              id:
                type: string
                format: uuid
              metadata: {}
              mime_type:
                type: string
              object_key:
                type: string
              original_name:
                type: string
              realm_id:
                type: string
                format: uuid
              size_bytes:
                type: integer
                format: int64
              updated_at:
                type: string
                format: date-time
              updated_by:
                type:
                  - string
                  - "null"
                format: uuid
              uploaded_by:
                type: string
                format: uuid
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
...